Points along Free-flowing rivers.

The points show projected population changes along rivers. Projections generated in TACAR (https://github.com/darrennorris/TACAR). The population projections are an updated and extended version of those published by Norris et. al. 2019 (https://doi.org/10.1016/j.biocon.2019.02.022).

location

South America.

coord. ref

WGS 84 (EPSG:3395).

Usage

points_bau_ffr

Format

points_bau_ffr is a dataframe with 160169 points and 15 fields, including:

BASIN_NAME

Name of major basin.

subbasn

Name of subbasn.

geom

sf geometry column.

Source

https://github.com/darrennorris/TACAR

Details

A lot of points along rivers.

References

Norris, D., et. al. (2019). Prospects for freshwater turtle population recovery are catalyzed by pan-Amazonian community-based management. Biological Conservation, 233, https://doi.org/10.1016/j.biocon.2019.02.022.

Grill, G., Lehner, B., Thieme, M. et al. (2019) Mapping the world’s free-flowing rivers. Nature 569, 215–221, https://doi.org/10.1038/s41586-019-1111-9.

Examples

if (FALSE) { # \dontrun{
# convert to sf object
sf_points_bau_ffr <- sf::st_as_sf(points_bau_ffr, crs = 3395)
} # }